+++ packages/portal-proto/src/index.ts
@@ -20,7 +20,7 @@ import FunctionButton from "@/components/FunctionButton";
 import EnumFacet from "@/features/facets/EnumFacet";
 import DateRangeFacet from "@/features/facets/DateRangeFacet";
 import ExactValueFacet from "@/features/facets/ExactValueFacet";
-import NumericRangeFacet from "@/features/facets/NumericRangeFacet";
+import NumericRangeFacet from "@/features/facets/NumericRangeFacet/NumericRangeFacet";
 import ToggleFacet from "@/features/facets/ToggleFacet";
 // hooks
 import { useIsDemoApp } from "@/hooks/useIsDemoApp";
